+++ precompile/contracts/juror/ioc_orders_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	hu "github.com/ava-labs/subnet-evm/hubbleutils"
-	"github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
+	"github.com/ava-labs/subnet-evm/orderbook"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 	"github.com/ethereum/go-ethereum/common"
 	gomock "github.com/golang/mock/gomock"
