+++ precompile/contracts/jurorv2/matching_validation.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	hu "github.com/ava-labs/subnet-evm/hubbleutils"
-	ob "github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
+	ob "github.com/ava-labs/subnet-evm/orderbook"
 	b "github.com/ava-labs/subnet-evm/precompile/contracts/bibliophile"
 	"github.com/ava-labs/subnet-evm/utils"
 	"github.com/ethereum/go-ethereum/common"
