+++ precompile/contracts/jurorv2/matching_validation_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 
 	hu "github.com/ava-labs/subnet-evm/hubbleutils"
-	ob "github.com/ava-labs/subnet-evm/plugin/evm/orderbook"
+	ob "github.com/ava-labs/subnet-evm/orderbook"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
